Implemented some modifications to allow support of more shader languages or binarys in the future like Cg or SPIR-V.
