+++ osu.Framework/Graphics/Shaders/ShaderSourceType.cs
@@ -1,0 +1,11 @@
+﻿// Copyright (c) 2007-2017 ppy Pty Ltd <contact@ppy.sh>.
+// Licensed under the MIT Licence - https://raw.githubusercontent.com/ppy/osu-framework/master/LICENCE
+
+namespace osu.Framework.Graphics.Shaders
+{
+    public enum ShaderSourceType
+    {
+        GLSL
+        //More will go here later like CG,SPIRV,etc.
+    }
+}
